Finished some initiative suggestions

+++ Cumulative1/Models/Course.cs
@@ -9,7 +9,7 @@ namespace Cumulative1.Models
     {
         public int ClassId { get; set; }
         public string ClassCode { get; set; }
-        public int TeacherID { get; set; }
+        public long TeacherID { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime FinishDate { get; set; }
         public string ClassName { get; set; }
